make: add common cflags to prevent optimizing out security checks
Flags updated based on Ben's comments. The new warn flag hit in a few
cases, to make it easy to corrleate they are included with this patch.

+++ lib/bdev/bdev.c
@@ -3527,8 +3527,8 @@ spdk_bdev_io_complete(struct spdk_bdev_io *bdev_io, enum spdk_bdev_io_status sta
 			 *  NOMEM_THRESHOLD_COUNT I/O to complete but for low queue
 			 *  depth channels we will instead wait for half to complete.
 			 */
-			shared_resource->nomem_threshold = spdk_max((int64_t)shared_resource->io_outstanding / 2,
-							   (int64_t)shared_resource->io_outstanding - NOMEM_THRESHOLD_COUNT);
+			shared_resource->nomem_threshold = spdk_max(shared_resource->io_outstanding / 2U,
+							   shared_resource->io_outstanding - NOMEM_THRESHOLD_COUNT);
 			return;
 		}
